+++ jib-core/src/integration-test/java/com/google/cloud/tools/jib/api/ContainerizerIntegrationTest.java
@@ -141,7 +141,7 @@ public class ContainerizerIntegrationTest {
     Assert.assertThat(history, CoreMatchers.containsString("bazel build ..."));
   }
 
-  private static void assertLayerSizer(int expected, String imageReference)
+  private static void assertLayerSize(int expected, String imageReference)
       throws IOException, InterruptedException {
     Command command =
         new Command("docker", "inspect", "-f", "{{join .RootFS.Layers \",\"}}", imageReference);
@@ -182,7 +182,7 @@ public class ContainerizerIntegrationTest {
     String imageReference = "localhost:5000/testimage:testtag";
     localRegistry.pull(imageReference);
     assertDockerInspect(imageReference);
-    assertLayerSizer(7, imageReference);
+    assertLayerSize(7, imageReference);
     Assert.assertEquals(
         "Hello, world. An argument.\n", new Command("docker", "run", "--rm", imageReference).run());
 
@@ -251,7 +251,7 @@ public class ContainerizerIntegrationTest {
     progressChecker.checkCompletion();
 
     assertDockerInspect("testdocker");
-    assertLayerSizer(7, "testdocker");
+    assertLayerSize(7, "testdocker");
     Assert.assertEquals(
         "Hello, world. An argument.\n", new Command("docker", "run", "--rm", "testdocker").run());
   }
@@ -293,7 +293,7 @@ public class ContainerizerIntegrationTest {
     progressChecker.checkCompletion();
 
     new Command("docker", "load", "--input", outputPath.toString()).run();
-    assertLayerSizer(7, "testtar");
+    assertLayerSize(7, "testtar");
     Assert.assertEquals(
         "Hello, world. An argument.\n", new Command("docker", "run", "--rm", "testtar").run());
   }