+++ jib-maven-plugin/src/test/java/com/google/cloud/tools/jib/maven/BuildImageMojoIntegrationTest.java
@@ -298,7 +298,7 @@ public class BuildImageMojoIntegrationTest {
             .trim());
   }
 
-  private static void assertLayerSizer(int expected, String imageReference)
+  private static void assertLayerSize(int expected, String imageReference)
       throws IOException, InterruptedException {
     Command command =
         new Command("docker", "inspect", "-f", "{{join .RootFS.Layers \",\"}}", imageReference);
@@ -367,7 +367,7 @@ public class BuildImageMojoIntegrationTest {
             new Command("docker", "inspect", "-f", "{{.Created}}", targetImage).run().trim());
     Assert.assertTrue(buildTime.isAfter(before) || buildTime.equals(before));
     assertWorkingDirectory("/home", targetImage);
-    assertLayerSizer(8, targetImage);
+    assertLayerSize(8, targetImage);
   }
 
   @Test
@@ -449,7 +449,7 @@ public class BuildImageMojoIntegrationTest {
     Assert.assertEquals(
         "Hello, world. An argument.\nrw-r--r--\nrw-r--r--\nfoo\ncat\nbaz\n",
         buildAndRun(simpleTestProject.getProjectRoot(), targetImage, "pom-extra-dirs.xml", false));
-    assertLayerSizer(9, targetImage); // one more than usual
+    assertLayerSize(9, targetImage); // one more than usual
   }
 
   @Test