+++ jib-gradle-plugin/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Can now containerize a JAR artifact instead of putting individual `.class` and resource files with `jib.containerizingMode = 'packaged'` ([#1760](https://github.com/GoogleContainerTools/jib/pull/1760/files))
+
 ### Changed
 
 ### Fixed